pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mercor-apex-finance-eval"
version = "0.1.0"
description = "Scaffold for public Mercor APEX finance evals with Reducto parsing and cost/value accounting."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "OpenAI"}]
dependencies = [
  "openai>=2.30.0",
  "python-dotenv>=1.0.1",
  "huggingface_hub>=0.25.0",
  "reductoai",
  "daytona-sdk>=0.158.0"
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[project.scripts]
apex-finance-eval = "mercor_apex_finance_eval.cli:main"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
mercor_apex_finance_eval = ["prompts/*.txt"]
